moved style components
